+++ cli/src/compile.js
@@ -39,26 +39,63 @@ export async function compile(serverinfo, dirpath, cleanzip = true) {
     // Generate the zip
     const zipPath = await createZip(dirpath);
 
+    utils.log(`Zip created: ${zipPath}`);
+
     // Transmit the zip
     return new Promise((resolve, reject) => {
-        http.get(commands.buildCommandUrl(serverinfo), (res) => {
+        const options = {
+            hostname: serverinfo.url,
+            port: serverinfo.port,
+            path: '/upload',
+            method: 'POST',
+            headers: {
+                'Content-Type': 'application/x-www-form-urlencoded',
+                'Content-Length': fs.statSync(zipPath).size
+            }
+        };
+
+        utils.log(`Initiating transmission to: ${commands.buildCommandUrl(serverinfo, commands.COMMAND_COMPILE)}`);
+
+        const clientRequest = http.request(options, res => {
+            utils.log(`STATUS: ${res.statusCode}`);
+            utils.log(`HEADERS: ${JSON.stringify(res.headers)}`);
+            res.setEncoding('utf8');
+
             let data = "";
-    
+
             res.on("data", (chunk) => {
                 data += chunk;
             });
-    
+
             res.on("end", () => {
                 utils.log(JSON.parse(data).explanation);
-                resolve("ok");
+
+                // Cleanup on finalize
+                if (cleanzip) {
+                    cleanZip(zipPath);
+                }
+
+                resolve("ok"); // Resolve only when receiving the response
             });
-        }).on("error", (err) => {
-            // Cleanup
+        });
+
+        clientRequest.on("error", (err) => {
+            // Cleanup upon error
             if (cleanzip) {
                 cleanZip(zipPath);
             }
 
             reject(err);
+        });
+
+        const zipFileStream = fs.createReadStream(zipPath);
+
+        zipFileStream.on("data", data => {
+            clientRequest.write(data);
+        });
+
+        zipFileStream.on("end", () => {
+            clientRequest.end();
         });
     });
 }
@@ -76,6 +113,10 @@ async function createZip(dirpath) {
 }
 
 function cleanZip(zipPath) {
+    if (!fs.existsSync(zipPath)) {
+        return;
+    }
+
     utils.deleteFile(zipPath);
     utils.log(`Zip file ${zipPath} deleted.`);
 }
